Introduce new LDAP configuration options
`auth.ldap.custom_base_dn` can be used as a substitute for the
hard-coded base DN used for both authentication and authorization. The
hardcoded version mandated to have the user from the connection to exist
in LDAP under the `uid=` attribute. The new base DN allows for fully
custom string. Three variables can be used and they will be substituted
before sending the DN to LDAP.
  - `${username_attr}` will be substitured to the value defined in
  `auth.ldap.username.attributetype`.
  - `${user}` is substituted to the username provided by the client.
  - `${device_dn}` is substituted to the string defined in
  `auth.ldap.device_dn`

Additionaly a typo was fixed in the `auth.ldap.device_dn` default
string.

`auth.ldap.filters.$num.{key|value|op}` can be used to specify
additional filter values compared to the predefined
`objectClass=Class`. At least the `key` and `value` must be present and
they represent an equality check. At each level (`$num`) an optional
operation can be defined (currently only "and" or "or"). This applies to
all the previously defined filters in a left-associative manner. There
is no way to express hierarchy. An example can be found in a comment in
the schema file. The default option is the old behaviour if no filter is
present in the config.

`auth.ldap.bind_as_user` must be set to `true` if a second bind
operation is required to authenticate the user provided in the
client. Otherwise the user will only used for authorization. If this
option is omitted or set to `false`, then the old behaviour is
used.